+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/ResponseMappings.kt
@@ -93,7 +93,7 @@ fun IdentifierTypeDb.toTypeKeyNameDto(): TypeKeyNameVerboseDto<String> {
 
 fun IdentifierTypeDb.toDto(): IdentifierTypeDto {
     return IdentifierTypeDto(technicalKey, businessPartnerType, name, abbreviation, transliteratedName, transliteratedAbbreviation,
-        details.map { IdentifierTypeDetailDto(it.countryCode, it.mandatory) })
+        format, categories.toSortedSet(), details.map { IdentifierTypeDetailDto(it.countryCode, it.mandatory) })
 }
 
 fun LegalFormDb.toDto(): LegalFormDto {